+++ src/constants.js
@@ -6,7 +6,11 @@ function define(name, value) {
 }
 
 const LOG = require('pino')({
-  level: process.env.LOG_LEVEL ? process.env.LOG_LEVEL : 'fatal',
+  level: process.env.LOG_LEVEL
+    ? process.env.LOG_LEVEL
+    : process.argv.includes('--debug')
+    ? 'debug'
+    : 'fatal',
   prettyPrint: true,
   name: 'app-sec'
 });
@@ -188,7 +192,25 @@ const resources = {
   BOT_PROTECTION_EXCEPTIONS:
     '/appsec/v1/configs/%s/versions/%s/security-policies/%s/transactional-endpoints/bot-protection-exceptions',
   BOT_ENDPOINT_COVERAGE_REPORTS: '/appsec/v1/bot-endpoint-coverage-report',
-  BOT_ENDPOINT_COVERAGE_REPORT: '/appsec/v1/configs/%s/versions/%s/bot-endpoint-coverage-report'
+  BOT_ENDPOINT_COVERAGE_REPORT: '/appsec/v1/configs/%s/versions/%s/bot-endpoint-coverage-report',
+  CUSTOM_CLIENT_SEQUENCE: '/appsec/v1/configs/%s/versions/%s/custom-client-sequence',
+  CONTENT_PROTECTION_RULES:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/content-protection-rules',
+  CONTENT_PROTECTION_RULE:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/content-protection-rules/%s',
+  CONTENT_PROTECTION_DETECTIONS: '/appsec/v1/content-protection-detections',
+  CONTENT_PROTECTION_RULE_DETECTION_SETTINGS:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/content-protection-rules/%s/detection-settings',
+  CONTENT_PROTECTION_RULE_SEQUENCE:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/content-protection-rule-sequence',
+  PENALTY_BOX_CONDITIONS:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/penalty-box/conditions',
+  PENALTY_BOX_CONDITIONS_ALL:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/penalty-box/conditions',
+  EVAL_PENALTY_BOX_CONDITIONS:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/%s/eval-penalty-box/conditions',
+  EVAL_PENALTY_BOX_CONDITIONS_ALL:
+    '/appsec/v1/configs/%s/versions/%s/security-policies/eval-penalty-box/conditions'
 };
 
 define('URIS', resources);